+++ packages/core/src/connectivity/scp-cf/resilience-options.ts
@@ -1,3 +1,5 @@
+import CircuitBreaker from 'opossum';
+
 export interface ResilienceOptions {
   /**
    * A boolean value that indicates whether to execute request to SCP-CF services using circuit breaker.
@@ -5,8 +7,9 @@ export interface ResilienceOptions {
   enableCircuitBreaker?: boolean;
 }
 
-export const circuitBreakerDefaultOptions = {
+export const circuitBreakerDefaultOptions: CircuitBreaker.Options = {
   timeout: 10000,
   errorThresholdPercentage: 50,
+  volumeThreshold: 10,
   resetTimeout: 30000
 };